+++ apps/swirl-docs/src/components/Layout/DocumentationLayout.tsx
@@ -1,9 +1,11 @@
-import { DocHeadline } from "@swirl/lib/docs/src/docs.model";
+import { DocHeadline, FrontMatter } from "@swirl/lib/docs/src/docs.model";
 import { NavItem } from "@swirl/lib/navigation";
 import { MDXRemote } from "next-mdx-remote";
 import { CategoryNav } from "./CategoryNav";
 import { DocLinksNav } from "./DocLinksNav";
 import Footer from "./Footer";
+import { DocumentationHeader } from "../Documentation/DocumentationHeader";
+import { ComponentPreview } from "../ComponentPreview";
 
 export type ComponentExample = {
   description: string;
@@ -11,18 +13,11 @@ export type ComponentExample = {
   title: string;
 };
 
-type FrontMatter = {
-  title: string;
-  description: string;
-  tags?: string[];
-  examples: ComponentExample[];
-};
-
 interface DocumentationLayoutProps {
   documentLinkList: DocHeadline[];
   categoryLinkList: NavItem[] | undefined;
-  mdxComponents: any;
   document: any;
+  mdxComponents?: any;
   frontMatter?: FrontMatter;
 }
 
@@ -33,6 +28,9 @@ export const DocumentationLayout = ({
   document,
   frontMatter,
 }: DocumentationLayoutProps) => {
+  const hasFrontMatterTitle = frontMatter?.title;
+  const isComponentDoc = frontMatter?.examples;
+
   return (
     <div className={`flex min-h-[calc(100vh_-_72px)]`}>
       <CategoryNav categoryLinkList={categoryLinkList} />
@@ -44,44 +42,12 @@ export const DocumentationLayout = ({
               className="flex flex-col justify-center items-center"
             >
               <article className="max-w-3xl px-4 mt-20">
-                {frontMatter?.title && (
-                  <header className="border-b-1 pb-12 mb-12">
-                    <div className="mb-1 inline-flex items-center">
-                      <h1 className="text-4xl text-text-default ">
-                        {frontMatter.title}
-                      </h1>
-                      {frontMatter.tags?.map((tag) => (
-                        <span
-                          className="bg-surface-neutral-subdued px-2 py-1 rounded-md ml-2 font-medium text-sm"
-                          key={tag}
-                        >
-                          {tag}
-                        </span>
-                      ))}
-                    </div>
-                    <p className="text-lg leading-line-height-xl text-text-default">
-                      {frontMatter.description}
-                    </p>
-                  </header>
+                {hasFrontMatterTitle && (
+                  <DocumentationHeader frontMatter={frontMatter} />
                 )}
-                {frontMatter?.examples?.map((example) => {
-                  return (
-                    <div key={example.url} className="mb-12">
-                      <h2>Variants</h2>
-                      <p className="text-lg text-text-default mb-12">
-                        This demo lets you preview the button component, its
-                        variations, and configuration options. Each tab displays
-                        a different type of button.
-                      </p>
-                      <iframe
-                        width="100%"
-                        height="500"
-                        src={example.url}
-                        frameBorder="0"
-                      ></iframe>
-                    </div>
-                  );
-                })}
+                {isComponentDoc && (
+                  <ComponentPreview frontMatter={frontMatter} />
+                )}
                 <MDXRemote {...document} components={mdxComponents} />
               </article>
             </main>