+++ apps/swirl-docs/src/pages/components/[componentDoc].tsx
@@ -7,6 +7,7 @@ import { createStaticPathsData } from "@swirl/lib/docs";
 import { createLinkLists } from "@swirl/lib/docs/src/links";
 import { ScriptProps } from "next/script";
 import { GetStaticProps } from "next";
+import { useEffect, useState } from "react";
 
 async function getComponentData(document: string) {
   return await generateMdxFromDocumentation(BASE_PATHS.COMPONENTS, document);
@@ -50,7 +51,11 @@ export default function Component({
   title: string;
   document: any;
 }) {
-  const components = {};
+  const [frontMatter, setFrontMatter] = useState<any>(null);
+
+  useEffect(() => {
+    setFrontMatter(document?.frontmatter);
+  }, [document]);
 
   return (
     <>
@@ -61,8 +66,7 @@ export default function Component({
         categoryLinkList={componentsNavItems}
         documentLinkList={[]}
         document={document}
-        mdxComponents={components}
-        frontMatter={document.frontmatter}
+        frontMatter={frontMatter}
       />
     </>
   );